Optimized message receive routine
The while loop conditional could be removed by moving the checksum check
outside the message parsing logic. The parse state counter could also be
incremented inside the loop, as opposed to setting it to the next state
during every step. All of this freed quite a bit of program space.