Added RAM Read/Write commands for pin control
With this added, it is possible to use the computer to toggle any
digital pin, drive any timer and pretty much anything else you wish to
do. The drawback is that it is slow; and thus not suitable for
serial/spi/lcd communication. More commands might be added for these in
particular.