WiFi: Turns out we didn't all this control
Burst size is unneeded and only adds TCP roundtrip times, and has
therefore been removed and code simplified. With ATE0 the command echo
logic can be turned off, which means we do not need to do any echo
response reading. As a result the logic became a whole lot simpler, the
SEND OK automatically handled by the receive routine and computer it is
connected to.